Extract each of the casper engine into seperate files
